Move patterns header file
