+++ diffkemp/patterns/diffkemp_patterns.h
